Generate missing Resize events by tracking window and resize dimensions
Track the dimensions of the last resize event and compare them to the
window dimensions each time an event is polled in order to determine
whether or not we need to generate new Resize events. We do this in
order to solve the problem where glutin (winit) is unable to produce
Resize events for Mac OS and possibly other operating systems.

The reason why we track the last emitted resize instead of the window
dimensions is so that we avoid generating duplicate events on systems
that already successfully generate Resize events.

Also cleans up the window's `size` and `draw_size` events by using the
glutin::Window methods specified for this purpose.