Remove Resize events generation now that upstream has been fixed.
tomaka/glutin#850 and tomaka/winit#118 provide a fix for missing
`Resized` events on OS X. The upstream fix is temporary until a proper
solution in relation to tomaka/winit#20 can be implemented. The temp fix
is basically the same as the workaround that we had here, so there
should no longer be any reason to maintain this code.